+++ src/index.js
@@ -1,5 +1,5 @@
-import * as movie from './js/homePage';
-import * as language from './js/languageSwitch';
+import * as homePage from './js/homePage';
+import * as languageIndex from './js/languageIndex';
 import * as trendTime from './js/trendTime';
 import * as pagination from './js/paginationPage';
 import * as theme from './js/themeSwitcher';
