+++ src/js/languageSwitch.js
@@ -1,7 +1,4 @@
 import { Movie } from './fetchMovie';
-import { fetchTrendAndMarkup, fetchSearchAndMarkup } from './fetchAndMarkup';
-import { trendMovie } from './homePage';
-import { keyword, keywordMovies } from './moviesKeyword';
 import { disabledEl, unlockEl } from './interfaceWork';
 
 export const keyLS = {
@@ -9,33 +6,10 @@ export const keyLS = {
   LS_GENRE_KEY_EN: 'themoviedb-genre-EN',
   LS_GENRE_KEY_UA: 'themoviedb-genre-UA',
 };
-const refs = {
+export const refs = {
   btnSwitchEN: document.querySelector('.switch-EN'),
   btnSwitchUA: document.querySelector('.switch-UA'),
 };
-
-refs.btnSwitchEN.addEventListener('click', onClickEN);
-refs.btnSwitchUA.addEventListener('click', onClickUA);
-
-function onClickEN() {
-  if (keyword === null) {
-    trendMovie.langCurrent = setLanguageToLS(Movie.language.ENGLISH);
-    fetchTrendAndMarkup(trendMovie);
-  } else {
-    keywordMovies.langCurrent = setLanguageToLS(Movie.language.ENGLISH);
-    fetchSearchAndMarkup(keywordMovies);
-  }
-}
-
-function onClickUA() {
-  if (keyword === null) {
-    trendMovie.langCurrent = setLanguageToLS(Movie.language.UKRAINIAN);
-    fetchTrendAndMarkup(trendMovie);
-  } else {
-    keywordMovies.langCurrent = setLanguageToLS(Movie.language.UKRAINIAN);
-    fetchSearchAndMarkup(keywordMovies);
-  }
-}
 
 export function setLanguageToLS(newLanguage) {
   localStorage.setItem(keyLS.LS_LANGUAGE_KEY, JSON.stringify(newLanguage));
