+++ src/library.js
@@ -1,2 +1,2 @@
 import * as theme from './js/themeSwitcher';
-//import * as language from './js/languageSwitch';
+import * as languageLibrary from './js/languageLibrary';
