+++ src/partials/header-library.html
@@ -1,33 +1,40 @@
 <header class="header header-mylib">
-    <div class="header-container container">
-        <div class="scroll-area"></div>
-        <nav class="header-nav animate__bounceInDown">
-            <a class="logo" href="/src/index.html" lang="en">
-                <svg class="logo__icon" width="24" height="24">
-                    <use href="/src/images/sprite.svg#film"></use>
-                </svg>
-                <span class="logo__name">Filmoteka</span>
-            </a>
-            <ul class="site-nav">
-                <li class="site-nav__item">
-                    <a href="/src/index.html" class="site-nav__link">Home</a>
-                </li>
-                <li class="site-nav__item">
-                    <a href="/src/library.html" class="site-nav__link site-nav__link--current">My library</a>
-                </li>
-            </ul>
-        </nav>
-        <ul class="filter-list animate__zoomIn">
-            <li class="filter-list__item">
-                <button type="button" class="filter-list__btn is-active">Watched</button>
-            </li>
-            <li class="filter-list__item">
-                <button type="button" class="filter-list__btn">Queue</button>
-            </li>
-        </ul>
-        <div class="language-switch animate__zoomIn">
+  <div class="header-container container">
+    <div class="scroll-area"></div>
+    <nav class="header-nav animate__bounceInDown">
+      <a class="logo" href="/src/index.html" lang="en">
+        <svg class="logo__icon" width="24" height="24">
+          <use href="/src/images/sprite.svg#film"></use>
+        </svg>
+        <span class="logo__name">Filmoteka</span>
+      </a>
+      <ul class="site-nav">
+        <li class="site-nav__item">
+          <a href="/src/index.html" class="site-nav__link">Home</a>
+        </li>
+        <li class="site-nav__item">
+          <a
+            href="/src/library.html"
+            class="site-nav__link site-nav__link--current"
+            >My library</a
+          >
+        </li>
+      </ul>
+    </nav>
+    <ul class="filter-list animate__zoomIn">
+      <li class="filter-list__item">
+        <button type="button" class="filter-list__btn is-active">
+          Watched
+        </button>
+      </li>
+      <li class="filter-list__item">
+        <button type="button" class="filter-list__btn">Queue</button>
+      </li>
+    </ul>
+    <!-- <div class="language-switch animate__zoomIn">
             <button class="switcher-EN">en</button>
             <button class="switcher-UA">ua</button>
-        </div>
-        </div>
+        </div> -->
+    <include src="./partials/languageSwitch.html"></include>
+  </div>
 </header>
