+++ src/partials/header-main.html
@@ -32,13 +32,16 @@
             </button>
         </form>
         <p class="warning-notification"></p>
-        <div class="trend-time-switch animate__zoomIn">
+        <!-- <div class="trend-time-switch animate__zoomIn">
             <button class="time-switch-day">day</button>
             <button class="time-switch-week">week</button>
         </div>
         <div class="language-switch animate__zoomIn">
             <button class="switch-EN">en</button>
             <button class="switch-UA">ua</button>
-        </div>
+        </div> -->
+        <include src="./partials/trendTime.html"></include>
+        <include src="./partials/languageSwitch.html"></include>
+    
     </div> 
 </header>
