+++ src/partials/languageSwitch.html
@@ -1,4 +1,4 @@
-<div class="language-switch">
+<div class="language-switch animate__zoomIn">
   <button class="switch-EN">en</button>
   <button class="switch-UA">ua</button>
 </div>
