+++ src/partials/trendTime.html
@@ -1,4 +1,4 @@
-<div class="trend-time-switch">
+<div class="trend-time-switch animate__zoomIn">
   <button class="time-switch-day">day</button>
   <button class="time-switch-week">week</button>
 </div>
